MAINT: Move contents of setmembership.py and refactor accordingly
