+++ src/porepy/fracs/meshing.py
@@ -356,7 +356,7 @@ def _tag_faces(grids, check_highest_dim=True):
         true_tip = np.intersect1d(may_be_tip, occurs_once)
 
         # Take the intersection between tip nodes and the nodes in this fracture.
-        _, local_true_tip = pp.utils.setmembership.ismember_rows(
+        _, local_true_tip = pp.array_operations.ismember_rows(
             true_tip, g_h.global_point_ind
         )
         tip_tag = np.zeros(g_h.num_nodes, dtype=bool)
@@ -366,7 +366,7 @@ def _tag_faces(grids, check_highest_dim=True):
         g_h.tags["node_is_fracture_tip"] = tip_tag
 
         on_any_tip = np.where(np.bincount(global_node_as_fracture_tip) > 0)[0]
-        _, local_any_tip = pp.utils.setmembership.ismember_rows(
+        _, local_any_tip = pp.array_operations.ismember_rows(
             on_any_tip, g_h.global_point_ind
         )
         tip_of_a_fracture = np.zeros_like(tip_tag)
@@ -492,7 +492,7 @@ def _assemble_mdg(
 
             # Find intersection between cell-node and face-nodes. Node nede to sort
             # along 0-axis, we know we've done that above.
-            is_mem, cell_2_face = pp.utils.setmembership.ismember_rows(
+            is_mem, cell_2_face = pp.array_operations.ismember_rows(
                 cn_all, fn, sort=False
             )
             # Now, for each lower-dimensional grid, either all of none of the cells
