+++ src/porepy/geometry/intersections.py
@@ -798,13 +798,13 @@ def polygons_3d(
 
                     # Uniquify the intersection points found on this segment of main.
                     # If more than one, the intersection is on the boundary of main.
-                    tmp_unique_isect, *rest = pp.utils.setmembership.uniquify_point_set(
+                    tmp_unique_isect, *rest = pp.array_operations.uniquify_point_set(
                         tmp_isect, tol=tol
                     )
                     if tmp_unique_isect.shape[1] > 1:
                         isect_on_boundary_main = True
 
-                isect, *rest = pp.utils.setmembership.uniquify_point_set(isect, tol=tol)
+                isect, *rest = pp.array_operations.uniquify_point_set(isect, tol=tol)
 
                 if isect.shape[1] == 0:
                     # The polygons share a plane, but no intersections
@@ -1002,14 +1002,14 @@ def polygons_3d(
                             isect = np.hstack((isect, loc_isect))
                             tmp_isect = np.hstack((tmp_isect, loc_isect))
 
-                    tmp_unique_isect, *rest = pp.utils.setmembership.uniquify_point_set(
+                    tmp_unique_isect, *rest = pp.array_operations.uniquify_point_set(
                         tmp_isect, tol=tol
                     )
 
                     if tmp_unique_isect.shape[1] > 1:
                         isect_on_boundary_other = True
 
-                isect, *rest = pp.utils.setmembership.uniquify_point_set(isect, tol=tol)
+                isect, *rest = pp.array_operations.uniquify_point_set(isect, tol=tol)
 
                 seg_vert_main_0 = (0, "not implemented for shared planes")
                 seg_vert_main_1 = (0, "not implemented for shared planes")
@@ -2335,7 +2335,7 @@ def split_intersecting_segments_2d(
         # NOTE: The tolerance used here is a bit sensitive, if set too loose, this
         # may merge non-intersecting fractures.
 
-        unique_all_pt, ia, ib = pp.utils.setmembership.uniquify_point_set(all_pt, tol)
+        unique_all_pt, ia, ib = pp.array_operations.uniquify_point_set(all_pt, tol)
 
         # Data structure for storing the split edges.
         new_edge = np.empty((e.shape[0], 0), dtype=int)
@@ -2375,7 +2375,7 @@ def split_intersecting_segments_2d(
         # Keep the old tags before uniquifying
         tags = new_edge[2:].copy().ravel()
         # Uniquify.
-        _, edge_map, all_2_unique = pp.utils.setmembership.unique_columns_tol(
+        _, edge_map, all_2_unique = pp.array_operations.unique_columns_tol(
             new_edge[:2].astype(int), tol
         )
         tag_info = (tags, all_2_unique)
@@ -2703,7 +2703,7 @@ def _intersect_pairs(p1: np.ndarray, p2: np.ndarray) -> np.ndarray:
         return np.empty((2, 0))
     else:
         # Do the intersection
-        _, ind = pp.utils.setmembership.ismember_rows(p1, p2)
+        _, ind = pp.array_operations.ismember_rows(p1, p2)
         pairs = p2[:, ind]
 
         # First sort the pairs themselves
